Agregue una alerta para palabras de longitud 0 o menores

+++ src/main/java/com/example/hanged/controller/WelcomeController.java
@@ -5,6 +5,7 @@ import com.example.hanged.view.GameStage;
 import com.example.hanged.view.WelcomeStage;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.HBox;
 
@@ -15,9 +16,25 @@ public class WelcomeController {
     private TextField secretWordTextField;
     @FXML
     public void onHandleButtonPlay(ActionEvent event) throws IOException {
-        SecretWord secretWord = new SecretWord(secretWordTextField.getText().trim());
-        GameStage.getInstance().getGameController().setSecretWord(secretWord);
-        WelcomeStage.deleteInstance();
-        System.out.println(secretWord.getWord());
+        if(secretWordTextField.getText().length()<=0) {
+            Alert alert = new Alert(Alert.AlertType.INFORMATION);
+
+            // Establecer el título de la alerta
+            alert.setTitle("Error");
+
+            // Establecer el encabezado de la alerta
+            alert.setHeaderText("Palabra demasiado corta");
+
+            // Establecer el contenido de la alerta
+            alert.setContentText("Por favor regrese e ingrese una palabra valida");
+
+            // Mostrar la alerta y esperar a que el usuario la cierre
+            alert.showAndWait();
+        }else {
+            SecretWord secretWord = new SecretWord(secretWordTextField.getText().trim());
+            GameStage.getInstance().getGameController().setSecretWord(secretWord);
+            WelcomeStage.deleteInstance();
+            System.out.println(secretWord.getWord());
+        }
     }
 }
